refactor: what an app's state means belongs to the protocol, not to a client
`@repo/app-operations` is what the cli and the dashboard reach the api through,
so the api depending on it pointed the wrong way — and through `@repo/api-client`
it pointed back at the api itself.

What the api wanted from it was never the client half. The status a row and a
release make together, the operation a state refuses and why, the hostname an
app is shown at, the account a release that never came up left behind: all of it
describes what the states in `@repo/protocol` mean, and none of it has ever
needed a request. So it moves there, where the cli, the dashboard and a tool can
each read it without going through one another.

The rest the api does for itself, as the routes already do: an environment edit
and a guest path are parsed against the schemas that define them.
