+++ apps/api/package.json
@@ -22,7 +22,6 @@
     "@ilbertt/better-auth-bun-sql": "^0.4.0",
     "@ilbertt/bun-sqlgen": "^0.5.0",
     "@modelcontextprotocol/server": "^2.0.0",
-    "@repo/app-operations": "workspace:*",
     "@repo/protocol": "workspace:*",
     "better-auth": "catalog:",
     "elysia": "^1.4.29",
