+++ packages/app-operations/src/release.ts
@@ -1,4 +1,3 @@
-import { ApiError } from '@repo/api-client/unwrap';
 import {
   HttpPortSchema,
   type TenantArguments,
@@ -43,24 +42,4 @@ export function configPatch({ args, port, extraPublicPort, environment }: Config
     ...(extraPublicPort !== undefined && { hasExtraPublicPort: extraPublicPort }),
     ...(environment !== undefined && { environment }),
   };
-}
-
-/**
- * The address to hand back, preferring a domain the owner brought: the platform hostname is what
- * nibrun issued, but a custom one is what they call their app.
- *
- * Active only. A brought domain is pending until the edge holds a certificate for it, and a link
- * that does not resolve yet is worse than the one that does.
- */
-export function servingHostname(
-  hostnames: ReadonlyArray<{ hostname: string; kind: string; state: string }>,
-): string {
-  const serving =
-    hostnames.find((entry) => entry.kind === 'custom' && entry.state === 'active') ??
-    hostnames.find((entry) => entry.kind === 'platform') ??
-    hostnames[0];
-  if (!serving) {
-    throw new ApiError('The app was created without a hostname.');
-  }
-  return serving.hostname;
 }
